Fixed the VineTrap :wine_glass:
Fixed the values for the vine trap, so it now slingshot the player
towards a maximum angle of 60 degrees (can still drag it down to 90
degrees).
So players always will be thrown forward and not so much upwards.
Plus ficed some issues with materials of some models
Started on the SpringTrap as well