Pickup "animation" and monguin texture cleanup
Changed Amber model; Why is amber_outside the inner mesh and
amber_inside the outer mesh?
Made the pickup "animation" (currently only on Sap) - there are two
types, one that flies towards camera and another that stays the same
distance from camera and flies out to the button in the canvas. Both
methods has the problem that we cannot draw 3D objects in front of
camera (or on the camera for that matter).
The "mouth (black line)" went wonky on the texture so it isn't there at
the moment, the eyes are a little off due to foldedness of the UVs ._.
Is it really intended that the monguin foot/lowerLeg should overlap at
times?
Updated material on Berries (herbs).
Repositioned the Health bar that hovers over the Monguin.